# api-docs/swagger.yaml
components:
  headers:
    CacheControl:
      schema:
        const: no-store
        type: string
    ContentTypeOptions:
      schema:
        const: nosniff
        type: string
    FrameOptions:
      schema:
        const: DENY
        type: string
...
    ReferrerPolicy:
      schema:
        const: strict-origin-when-cross-origin
        type: string
...
    Vary:
      description: Response selection variance.
      schema:
        type: string
...
    XRequestID:
      description: Selected immutable request correlation value.
      schema:
        maxLength: 128
        minLength: 1
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$
        type: string
  parameters:
...
    XRequestID:
      description: A missing, invalid, repeated, or comma-combined candidate is replaced
        with a generated request ID.
      in: header
      name: X-Request-ID
      required: false
      schema:
        maxLength: 128
        minLength: 1
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$
        type: string
  schemas:
    BoundedName:
      maxLength: 100
      minLength: 1
      pattern: ^(?![\u0009-\u000D\u0020\u0085\u00A0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000])(?!.*[\u0009-\u000D\u0020\u0085\u00A0\u1680\u2000-\u200A\u2028\u2029\u202F\u205F\u3000]$)[^\u0000-\u001F\u007F-\u009F]*$
      type: string
...
    ErrorDetail:
      additionalProperties: false
      properties:
        detail:
          maxLength: 200
          minLength: 1
          type: string
        source:
          $ref: '#/components/schemas/ErrorSource'
      required:
      - detail
      type: object
    ErrorSource:
      oneOf:
      - additionalProperties: false
        properties:
          pointer:
            maxLength: 256
            minLength: 1
            type: string
        required:
        - pointer
        type: object
      - additionalProperties: false
        properties:
          parameter:
            maxLength: 256
            minLength: 1
            type: string
        required:
        - parameter
        type: object
      - additionalProperties: false
        properties:
          header:
            maxLength: 256
            minLength: 1
            type: string
        required:
        - header
        type: object
...
    Greeting:
      additionalProperties: false
      properties:
        message:
          type: string
      required:
      - message
      type: object
...
    HelloCreate:
      additionalProperties: false
      properties:
        name:
          $ref: '#/components/schemas/BoundedName'
      required:
      - name
      type: object
...
    Problem_internal_error:
      additionalProperties: false
      properties:
        code:
          const: internal_error
          type: string
        detail:
          const: Internal server error
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 500
          type: integer
        title:
          const: Internal Server Error
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
    Problem_invalid_request:
      additionalProperties: false
      properties:
        code:
          const: invalid_request
          type: string
        detail:
          const: Request is malformed
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 400
          type: integer
        title:
          const: Bad Request
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
    Problem_not_acceptable:
      additionalProperties: false
      properties:
        code:
          const: not_acceptable
          type: string
        detail:
          const: No acceptable response representation is available
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 406
          type: integer
        title:
          const: Not Acceptable
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
...
    Problem_payload_too_large:
      additionalProperties: false
      properties:
        code:
          const: payload_too_large
          type: string
        detail:
          const: Request body is too large
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 413
          type: integer
        title:
          const: Content Too Large
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
...
    Problem_unsupported_media_type:
      additionalProperties: false
      properties:
        code:
          const: unsupported_media_type
          type: string
        detail:
          const: Request representation is not supported
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 415
          type: integer
        title:
          const: Unsupported Media Type
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
    Problem_validation_failed:
      additionalProperties: false
      properties:
        code:
          const: validation_failed
          type: string
        detail:
          const: Request validation failed
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 422
          type: integer
        title:
          const: Unprocessable Content
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
...
  securitySchemes:
    BearerAuth:
      bearerFormat: Firebase ID token
      description: Firebase ID token verified with revocation checking.
      scheme: bearer
      type: http
info:
  description: Portable REST contract implemented with Echo 5.3.
  title: Echo Playground Portable API
  version: 1.0.0
...
openapi: 3.1.2
paths:
...
  /v1/hello:
    get:
      description: The query string is closed; unknown, repeated, or malformed parameters
        are rejected.
      operationId: getHello
      parameters:
      - $ref: '#/components/parameters/XRequestID'
      responses:
        "200":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Greeting'
            application/json:
              schema:
                $ref: '#/components/schemas/Greeting'
          description: Successful response
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "400":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_invalid_request'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_invalid_request'
          description: Controlled invalid_request failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "406":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_not_acceptable'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_not_acceptable'
          description: Controlled not_acceptable failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "500":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_internal_error'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_internal_error'
          description: Controlled internal_error failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
      security: []
      summary: Get the default greeting
    post:
      description: The query string is closed; unknown, repeated, or malformed parameters
        are rejected.
      operationId: createHello
      parameters:
      - $ref: '#/components/parameters/XRequestID'
      requestBody:
        content:
          application/cbor:
            schema:
              $ref: '#/components/schemas/HelloCreate'
          application/json:
            schema:
              $ref: '#/components/schemas/HelloCreate'
        required: true
      responses:
        "200":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Greeting'
            application/json:
              schema:
                $ref: '#/components/schemas/Greeting'
          description: Successful response
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "400":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_invalid_request'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_invalid_request'
          description: Controlled invalid_request failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "406":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_not_acceptable'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_not_acceptable'
          description: Controlled not_acceptable failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "413":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_payload_too_large'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_payload_too_large'
          description: Controlled payload_too_large failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "415":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_unsupported_media_type'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_unsupported_media_type'
          description: Controlled unsupported_media_type failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "422":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_validation_failed'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_validation_failed'
          description: Controlled validation_failed failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "500":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_internal_error'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_internal_error'
          description: Controlled internal_error failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
      security: []
      summary: Create a personalized greeting
...